+++ src/ast/nodes/shared/FunctionNode.ts
@@ -5,6 +5,7 @@ import FunctionScope from '../../scopes/FunctionScope';
 import type { ObjectPath, PathTracker } from '../../utils/PathTracker';
 import type BlockStatement from '../BlockStatement';
 import Identifier, { type IdentifierWithVariable } from '../Identifier';
+import type { ExpressionEntity } from './Expression';
 import { UNKNOWN_EXPRESSION } from './Expression';
 import FunctionBase from './FunctionBase';
 import { type IncludeChildren } from './Node';
@@ -93,6 +94,10 @@ export default class FunctionNode extends FunctionBase {
 	initialise(): void {
 		super.initialise();
 		this.id?.declare('function', this);
+	}
+
+	protected addArgumentToBeDeoptimized(argument: ExpressionEntity) {
+		this.scope.argumentsVariable.addArgumentToBeDeoptimized(argument);
 	}
 
 	protected getObjectEntity(): ObjectEntity {